feat: move table group package
