+++ src/main/java/kitchenpos/application/dto/TableGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.application.dto;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
