+++ src/main/java/kitchenpos/domain/tablegroup/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.tablegroup;
 
 import kitchenpos.domain.table.OrderTable;
 import org.springframework.util.CollectionUtils;
