+++ src/main/java/kitchenpos/domain/tablegroup/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.tablegroup;
 
 import org.springframework.data.repository.Repository;
