+++ src/test/java/kitchenpos/fixture/TableGroupFixture.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
 import kitchenpos.application.dto.TableGroupRequest;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
